chore(exec): post-1238 cleanup -- shell_write docs, test location, visibility (#1240)
