+++ AGENTS.md
@@ -85,7 +85,7 @@ Canonical parameter lists live in the `types` module (`crates/aptu-coder-core/sr
 - `git_ref` is supported on both `analyze_directory` and `analyze_symbol` to restrict analysis to files changed relative to a git ref.
 - `working_dir` on `edit_overwrite` and `edit_replace` sets the base directory for path resolution (default: server CWD). The resolved target path must be within working_dir.
 - `edit_replace` accepts empty `new_text` to delete the matched block. CRLF line endings in `old_text` are normalized to LF before matching. A stale-context circuit breaker fires after 5 consecutive `not_found` or `ambiguous` failures on the same (session_id, path) pair, returning a directive error; the map is capped at 1024 entries.
-- `exec_command` accepts an optional `timeout_secs` (integer >= 0; 0 or omitted means no limit). When the child process exceeds the limit it is killed; `timed_out: true` is set in the response and `exit_code` is null. Heredoc syntax with a missing closing delimiter is rejected before spawn.
+- `exec_command` accepts an optional `timeout_secs` (integer >= 0; 0 or omitted means no limit). When the child process exceeds the limit it is killed; `timed_out: true` is set in the response and `exit_code` is null. Heredoc syntax is rejected before spawn in four cases: file-write pattern (cat/tee/redirect + `<<`), stdin-consuming flag (`--body-file -`, `--data -`, etc.) paired with `<<`, `params.stdin` set while command contains `<<`, or missing closing delimiter.
 - `exec_command` accepts an optional `drain_timeout_secs` (integer >= 0; 0 or omitted means 500ms default, negative = INVALID_PARAMS, positive = drain window in milliseconds). Controls how long the post-exit drain waits for a background subprocess holding pipes open before returning `output_truncated: true`.
 - `analyze_symbol` uses an L2 on-disk call-graph cache in addition to the L1 in-memory LRU. Configure with `APTU_CODER_DISK_CACHE_DIR` (default: `$XDG_DATA_HOME/aptu-coder/analysis-cache`); disable with `APTU_CODER_DISK_CACHE_DISABLED=1`.
 - `call_frequency` on `analyze_symbol` is filtered out when the `Functions` field is not in the projected fields set.
